+++ projects/ngx-material/src/components/index.ts
@@ -4,3 +4,4 @@ export * from './header/header.component';
 export * from './divider/ngx-divider.component';
 export * from './icon-button/icon-button.component';
 export * from './icon-anchor/icon-anchor.component';
+export * from './chip/chip.component';